Rename http sync_client and async_client to sync and aio
