+++ core/http/sync.py
