+++ main/refbooks/downloaders/url_csv.py
@@ -12,7 +12,7 @@ import gzip
 from io import StringIO
 
 # NOC modules
-from noc.core.http.sync_client import HttpClient
+from noc.core.http.sync import HttpClient
 from noc.core.comp import smart_text
 from .base import BaseDownloader
