+++ sa/profiles/Dahua/DH/middleware/dahuaauth.py
@@ -12,7 +12,7 @@ import codecs
 
 # NOC modules
 from noc.core.script.http.middleware.base import BaseMiddleware
-from noc.core.http.sync_client import HttpClient
+from noc.core.http.sync import HttpClient
 from noc.core.comp import smart_bytes
