+++ sa/profiles/IRE-Polus/Horizon/middleware/horizonauth.py
@@ -10,7 +10,7 @@ import orjson
 
 # NOC modules
 from noc.core.script.http.middleware.base import BaseMiddleware
-from noc.core.http.sync_client import HttpClient
+from noc.core.http.sync import HttpClient
 
 
 class HorizonAuthMiddeware(BaseMiddleware):
